Moved out VdevLabel
